extract InsertionDetourTimeCalculator from InsertionCostCalculator
